NeuralNetBase class in nn_util handling generic policy/value things
